renamed nativeScrolling.ts test to scrollToTIme, replace material theme cell tests to smoothCellLines.ts file
