+++ testing/testcafe/tests/scheduler/grouping/etalons/material-theme-scrolling-vertical.png
